requirements.txt
#geminiai
kubernetes
requests
slack-sdk
streamlit
typer[all]
pyyaml
pytest
scikit-learn
joblib
matplotlib